+++ src/main/java/by/itech/lab/supplier/domain/ItemsInApplication.java
@@ -1,8 +1,16 @@
 package by.itech.lab.supplier.domain;
 
+import lombok.AllArgsConstructor;
+import lombok.Builder;
+import lombok.Data;
+import lombok.NoArgsConstructor;
+
 import javax.persistence.Entity;
 import javax.persistence.Table;
 
+@Data
+@AllArgsConstructor
+@Builder
 @Entity
 @Table(name = "items_in_application")
 public class ItemsInApplication {
