+++ src/main/java/by/itech/lab/supplier/domain/ItemsInWarehouse.java
@@ -1,8 +1,16 @@
 package by.itech.lab.supplier.domain;
 
+import lombok.AllArgsConstructor;
+import lombok.Builder;
+import lombok.Data;
+import lombok.NoArgsConstructor;
+
 import javax.persistence.*;
-import java.util.HashSet;
-import java.util.Set;
+
+@Data
+@AllArgsConstructor
+@NoArgsConstructor
+@Builder
 @Entity
 @Table(name = "items_in_warehouse")
 public class ItemsInWarehouse {
