+++ src/main/java/by/itech/lab/supplier/domain/Retailer.java
@@ -1,8 +1,17 @@
 package by.itech.lab.supplier.domain;
 
+import lombok.AllArgsConstructor;
+import lombok.Builder;
+import lombok.Data;
+import lombok.NoArgsConstructor;
+
 import javax.persistence.*;
 import java.util.Objects;
 
+@Data
+@AllArgsConstructor
+@NoArgsConstructor
+@Builder
 @Entity
 @Table(name = "retailer")
 public class Retailer {
@@ -49,15 +58,4 @@ public class Retailer {
         this.retailersCol = reailersCol;
     }
 
-
-
-    @Override
-    public String toString() {
-        return "Retailer{" +
-                "id=" + id +
-                ", fullName='" + fullName + '\'' +
-                ", identifier='" + identifier + '\'' +
-                ", retailersCol='" + retailersCol + '\'' +
-                '}';
-    }
 }
