+++ src/main/java/by/itech/lab/supplier/domain/Zone.java
@@ -1,10 +1,18 @@
 package by.itech.lab.supplier.domain;
 
+import lombok.AllArgsConstructor;
+import lombok.Builder;
+import lombok.Data;
+import lombok.NoArgsConstructor;
+
 import javax.persistence.*;
 import java.util.HashSet;
-import java.util.Objects;
 import java.util.Set;
 
+@Data
+@AllArgsConstructor
+@NoArgsConstructor
+@Builder
 @Entity
 @Table(name = "zone")
 public class Zone {
@@ -40,14 +48,4 @@ public class Zone {
         this.taxes = taxes;
     }
 
-
-
-    @Override
-    public String toString() {
-        return "Zone{" +
-                "id=" + id +
-                ", zone='" + zone + '\'' +
-                ", taxes=" + taxes +
-                '}';
-    }
 }
